Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
motor==3.3.2
beanie==1.23.6
pydantic==2.5.2
numpy==1.26.2
scikit-learn==1.3.2
python-decouple==3.8
pydantic-settings==2.1.0
email-validator==2.1.0
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.6